Fix swapped assert (#8)
* Fix swapped assert

Test was checking for a string within the XML node that could not
possibly exist as it was never set (possibly an old default within the
system that was removed). It was also incorrectly asserting that the 
actual stored value was NOT there. This should be swapped, and 
then the inverse check can be safely removed.