Swap order of date picker operations (#9)
* Swap order of date picker operations

Currently, date picker click operations selects month before year. This
is an issue when the until date is the following year (I.E if the test
is run on New Year's Eve) as selecting an earlier month of the current
year is disabled. Selecting the year first ensures that this does not
occur.

* Change four spaces to tabs

For consistency with the rest of the file.