Feature/13 embeddedtester (#14)
* Update IntegTester libs and embed, upgrade to sbt 1.1.1, fix build scripts
* Fix warning and changes ps deps
* Do yarn install as part of it
* Make sure we don't export jars
* Use runner instead of runmain
* Cache ps compilation, remove warnings
* Bring back restrictions for parallel
* Run scalacheck tests in new jvm
* Explain the support server a bit more
* Ensure locale is set correctly in headless Chrome

There is a known issue in Chrome testing in which the default locale
is set differently in headless and headful modes. This causes issues in
autotests (for example, the ModerationProgressTest) when using headless
mode. This fix accounts for the problem.

* Update README.md
* Change default suitename in application.conf.example

Also, put a small explanation in the README.md detailing the importance
of using testng-codebuild.yaml.