Add develop branch and document new git workflow
It was discussed in the Dev mailing list that we'll move to a workflow
more in line with Git Flow. To support this, we need to add a develop
branch - the main thing we were lacking.

Added details in the CONTRIBUTING.md, and also did a little bit of
reformatting (mainly due to long lines).