Add undo button, rework button disabling
"Save" and "Revert Changes" buttons are now only enabled when the
contents of the textfield differ from the property currently saved in
the database. "Delete" is now only enabled when there is a current
notice set in the database, as opposed to just when the textfield
contains something as before.