Travis: Capture artefacts from test runs (and refactor S3 commands)
We need to outputs from the test stage no matter what (after_script) so
that if things do fail, we can readily see what has occurred.

Also, we needed to make things a lot more DRY - so this is the first
step of that. (Will be interesting to see how Travis interpolates the
variables.