Travis: Manually specify permutations to run TestNG with New UI
Attempt with build matrix failed - incorrect usage, so now we specify
manually.

Also need to make sure that we have individual folders on S3 for each
run so as not to clobber results between legacy and new UI.

And while working on this noted that we're not saving screenshots, so
added support for this.

All that highlighted the need for refactoring the scalacheck tests jobs
to make them more DRY.