Travis: Disable functional tests
Today the version of Chromium being installed on bionic has moved from
78 to 79. This has caused problems with many of the tests, however this
is right when we were planning to release.

Due to tests still running on the Edalex CI server and that they all ran
fine yesterday (on Chromium 78) we're happy to disable for now to
finalise the release, and then we will work to address the issue.