Travis: Pre-build IntegTester
A few minutes are spent as part of every functional test job on travis
building the integ tester. So by pre-building we should be able to save
some significant time for the whole build - maybe even an hour.

(Note: Locally it only takes < 1 minute to build the IntegTester, but
Travis machines have far fewer resources and so take significantly
longer.)