CI: Add GitHub actions trigger for component/** branches
For large features, we build them on a component branch which is then
merged back on feature completion - and various sub branches come off
them as each part of the feature is developed. As a result, we should
also add this to the list of branches to build both on merge (push)
and on PRs targeting them.