Remove @JsonIgnore from 'dateCreated' and 'dateModified';
In order to make the DAO extend GenericInstitutionDao, this entity needs
full information of an institution. So replace 'institutionId' with an
instance of 'Institution'. The relation should be many-to-one. And when
exporting an institution, there is no need to export the institution
information together with this entity so add '@XStreamOmitField'.