Property setup lazy loading for RichTextEditor
The RichTextEditor pulls in the whole of TinyMCE which has some code
that will execute on import. This code was breaking the new Jest test as
it was expecting various values to be pulled from the `window` object.

Lazily loading, helps with both of these.

However, this is not splitting the bundle, as parcel does not support
this form of code splitting.