Add initial shell of a JS oEQ REST API library
The plan being that this will provide a level of decoupling for the New UI
work. This will allow for easier testing of this interface, and then it
can be mocked in JS tests testing the New UI. Further, it could possibly
be used for any other JS projects.

Currently, it only has support for the basic login/logout methods, and
getting user details. However it will have more added shortly, and then be
used from the current New UI work.

Once that is shown to be working, we can split off into it's own repository.