chore(ci): Prepackage coverage report
The uploading of the coverage reports was taking a very long time and
sometimes failing. I believe this is because the artefact upload action
wasn't really designed for uploading directories with tens of thousands
of files.

Considering the upload of the primary artefacts only takes a few minutes
(and is a few hundred megabytes), I'm assuming this will save us
significant time and shorten our test runs by about 50%.