Fix structured selection session course panel display overflow issues
The main issue here was that <div id=col2> contains rouge whitespace. I couldn't find exactly what code was inserting that whitespace, so the lest risky option is to just alter the minimum width. I also altered some of the text size in the course selection panel so more information be displayed.