Fix missing icons in RichTextEditor
It seems that from TinyMCE (which we upgraded to recently) you needed to
explicitly include the icons. This can be done simply with an import
such as:

    import 'tinymce/icons/default';

But while at it I noticed all the TinyMCE imports had been done CommonJS
style rather then ES2015, so I updated them all for consistency.