Restyle new search bar
So as to be a bit cleaner, and similar to a number of Material Design
search bars you see about the place - so as to hopefully feel
contemporary to users.

Required custom styling, but I think that's acceptable in this case as
we wanted something a little different and it is a key/special component
as one of the main ones users interact with.