fix(storybook): Make RichEditor stories work
I'm not convinced this ever worked. First I thought it was just missing the renderData which oEQ puts in the header of the pages. However, adding this unveiled that the TinyMCE component failed to download it's stylesheets due to it attempting to determine the URL from <base>. Unfortunately Storybook also sets a <base> element and so I couldn't replace that via preview-head or manager-head. The end result, was to modify the component to optionally support receiving the URL via props.

#2024