chore: Remove dead TS code
There was a good amount of code left over from some inital approaches to how the data layer would be done for the New UI. We've current headed down a different path and so this code was not being used.

Also of notable mention is that we're not using Redux, so those dependencies have been removed. However I think there is still further work to be done here - might be some left over deps.