chore: Add typechecking to dev:build
Parcel.js uses babel for it's TS support and as a result does not also do the Typescript typechecking. However when refactoring or doing general dev it's common to use the `dev:build` script, so it'd be good to have it quickly check the types for us (via `tsc --noEmit`) rather than having to do it manually or wait for the CI.

Sure a lot of dev is also just done with the `dev` script, however that's using `parcel watch` so it doesn't make sense to add it there. (As it'd only be valid at the first run - not the incremental builds.)