chore(ci): Revert GHA checkout action to v1
There is currently an issue we're having with GHA where the wrong commit
is built. This appears to be related to issue
https://github.com/actions/checkout/issues/237

There is a comment there from 4 days ago
(https://github.com/actions/checkout/issues/237#issuecomment-662662682)
where the issue is now acknowledged but it's noted as a complex issue
with no definite date for resolution.

Others have reverted to v1, so it seems worth trying - especially if the
availability of the fix seems to be an unknown (possibly long) time
away.