Do not redirect if href is undefined
If the response of a Legacy content submit is neither a PageContent
nor a ChangeRoute, a redirect is performed even though where the page
should be redirected is not provided. This results in the page being
redirected to 'undefined'. So this fix ensures that not redirect is
performed if href is undefined.