Make the end of date range inclusive in searches.
Firstly, change the time of start to '00:00:00' and that of end to
'23:59:59'. Secondly, to ensure the old Search page keeps unchanged, add a
new boolean variable to flag if a search is performed by 'Search2' API.
And then in the function that prepares Lucine queries, use proper timezone
based on the flag.