front-end: Change OEQ.Common.ItemStatus to a string literal union
It seems the enum is TS has issues in Jest land. And although it seems like a great language feature, it's really rather half baked - it's not a full enum as you'd find in any modern language.

So it seems that's why you see the string literal unions used in other TS/JS libraries; and it made sense for us to follow.

However there was still a need to be able to iterate over the values of the resulting type, so also brought in runtypes to support this.