fix(front-end): Properly handle when there are _no_ attachments
When there we no attachments the SearchResult was getting in an infinite loop of state updates. Although it was happy when with an empty collection ([]), for some reason (even though there is a default param to []) when it 'attachments' was undefined, things spiraled out of control.

This is the case when a user has insufficient privileges to see attachments - and then the property is omitted from the search2 response.