chore(front-end): Refactor viewer determination in ItemAttachmentLink
The initial implementation morphed overtime, and what was left was really for the future and doco. Review feedback highlighted in its current state (and honoring YAGNI) it could be simplified into a ternary. This in term has also helped to make it more declarative, so that's a win. :)

Feedback from PR #2411