security(server): Re-use Tomcat basedir on service restarts
If the Tomcat service is restart, historically oEQ used
the same basedir. To ensure compatibility with tests, we
do the same here.

This is part of the fix for GHSA-79w4-xjfh-9rmf.

(This was a manual backport, and included minor fixes for the logging
compared to the other stable branches for 2018.x, 2019.x, 2020.x.)