Support accessing Legacy Search page through '/access/selectoradd/searching.do'.
The intention of this commit is to differentiate the URL
used to access the old Search UI from Selection Session.
With the changes made in #2414, once New Search UI is enabled,
'searching.do' will be captured by react-route, which results
in incorrect contents of Search page displayed in Selection
Session. Therefore, we need to have a new URL to access old
Search page in Selection Session.