Navigate from new Search page to ItemSummary page in Selection Session.
To display the Selection Session specific ItemSummary
page, extra information such as the ID of a Selection
Session and the Item count of a Search result is required
to help the legacy server rendering. So we can collect
Selection Session info from 'renderData' and pass Search
result data from component 'SearchPage' to 'SearchResult',
and then add query params to each Item title link.

Please note that MUI Link is used for the navigation in
Selection Session. This is due to the complexity of creating
a route in this context. Once we have a new component for
ItemSummary page, we should come back and use React Router
Link.