bugfix(front-end): Improve SearchPage error handling
A number of possible errors were not being caught. This would result in
spurious console log entries, and lack of feedback to the user that an
issue had occurred.

Most of these are now addressed, by simply shutting down oEQ and then
chasing things popping up in the console log. Or in a few cases, just by
 manual code analysis and looking at the user result.