bugfix(front-end): Stop double render on leaving a legacy page
This required essentially an inversion of components and the passing of
data. But more so it was a complete tangle before with data and function
calls going in multiple directions - so strictly speaking I can't say it
 was just an inversion.

But the crux here was to make it that LegacyContent is not only
concerned with fetching legacy content, but also the rendering of it.
Hence why the core part of `LegacyPage` and then the `render()` passed
in from `IndexPage` was merged in.

I did attempt to tread lightly and only do what was needed. (Although I
may have removed a superfluous React.memo here and there.)

#2262