refactor(front-end): Manage execution of scripts in LegacyContentRenderer
There was some complexity in timing of when these legacy scripts
returned from the server should be executed. However the responsibility
was a bit split across components. When looking closer it also
highlighted that some of the legacy content rendering which was shifted
to `LegacyContent` should be in `LegacyContentRender`.

These changes also seem to address a good number of tests which were
previously failing.