bugfix(front-end): Default to NO for fullscreenMode
With the previous default of YES there was a mismatch with what the
server defaults to, so on the first LegacyContent request the front-end
would trigger an update (and re-render) of the Template. This was done
with the call to updateTemplate with the `contents` from the request.

Unfortunately Template.tsx uses a ternary expression to build the layout
of the page, and this means if fullscreenMode changes the LegacyContent
component is unmounted and then recreated - which with the new work
triggers a double request to the LegacyContent API.

If this solution is not suitable - i.e. if there are different issues in
 the future - then we may need to look at pushing the legacy content out
into a Context.