refactor(front-end): Force render of LegacyContent on updated Page Content
Legacy oEQ is used to being served up and re-rendered by the browser for
new pages. However if a react app simply changes the content of a
component/DOM only slightly, then React will optimise the rendering and
only change what is needed.

This is a problem where in the background some AJAX stuff has modified
the DOM - say with a progress bar - which traditionally would simply be
removed as a side effect of a full page re-render (due to server side
rendering). However if in React we ask to update the component with the
same content from the server, then they end up hanging around - so a
100% complete progress bar just sits there.

By clearing `content` with `setContent(undefined)` in `LegacyContent`,
that triggers removing that part of the DOM tree and instead showing
the spinner while updates commence. Typically though, this is super
brief as the request has already been made.

It wasn't done at the lowest level of wrapping the request, as that same
logic is shared when there are AJAX and user update requests - not only
full page updates.