refactor(front-end): Minimise unmounting of Template children
Due to the ternary expression for the `layout` in `Template`, when the
fullscreen mode was changed all the `children` would be unmounted. This
was fine in the majority of cases, however for `LegacyContent` pages
this mode change and unmounting (and subsequent remounting) triggered
duplicate requests to the server. Which in turn could result in
incorrect data and server/session state.

For example, when visiting a report which had not yet been cached, the
first request would start the generation and then the second would get
returned a path for a HTML file which was empty. The browser would then
request this and cache it. So until the user pressed CTRL-F5 they'd see
no report.

But that example just demonstrates the potential for greater issues -
e.g. there possibly be an issue in the context of auditing and view
counts.

So although the use of the ternary expression is possibly more
idiomatic, in this case we needed to do something slightly different.
Possibly when we no longer have `LegacyContent` we could return to a
ternary expression to reduce the DOM size - but probably only a minor
consideration.