refactor(front-end): useReducer with SearchPage
This was done on a basic level to address issues with React hook
dependency lists. But more importantly so as to improve readability of
the code, as all the states and the chained side effects had reached a
point of being almost incomprehensible.

The hope this will be easier to maintain, and make it easier as we add
additional search features in the not too distant future.