Invalidate all cookies
I found that invalidating just the url key cookie cache only works
sometimes. If you continue to spam click you can still get this out of whack.
Possibly because the are based only on URLs. Once the fills up it makes new entries.
So invalidating the whole cache for this user seems to work better.