Update to @tinymce/tinymce-react 3.8.4
Attempted to upgrade to latest (3.12.2) but that version requires
a peer dependency of React 17. Since we have not already migrated
React to 17, we can not use latest tinymce-react. I tested it, and it completely breaks image upload.
3.8.4 is the latest version that still uses React 16, and I have confirmed it works.
There was a slight type change needed for the BlobInfo interface, but aside from that everything
appears to be in order.