chore(codebuild): Use sbt-extras wrapper
The build is failing on CodeBuild due to memory issues. This is strange
as there were none on GitHub Actions (GHA). So the only difference is
that on
GHA we use the sbt-extras wrapper where as on CodeBuild we use an
installed SBT from an APT repo from when the docker image was built 1-2
years ago.

Hopefully this helps.

#1691