chore(SBT): Address IO.zip deprecation
With SBT's move to binary reproducible builds IO.zip has now been
replaced with one which takes an Option[Long] for timestamp. I believe
the closest implementation is to provide a timestamp for the build
rather than None - which would be what you'd do if you did want binary
compatible builds (but we'd also have to address everything else).