refactor(server): Replace decompiled MD5 implementation
This class (Md5) was the source of the most warnings in the CheckStyle
report. When looking at the class it became obvious it was just a
decompiled MD5 implementation which is rather redundant when there are
many standard offerings out there now. Indeed, even across the oEQ base
although there are many uses of this, there are also uses of others.

So to at least tidy things up, this has been refactored. However
arguably a refactoring to consolidate hashing on a broader scale should
be undertaken. This might be easier now with this one tidied up.

(Added the unit test before the refactor so that I could validate there
was no change.)