chore(deps): Rollback Apache Curator to 4.3.0
Unfortunately 5.1.0 requires a newer version of Zk and that version is
not available as a standard debian package yet - so we'd have to change
our CD Zk cluster to test with. But at least 4.3.0 is still
significantly newer that what we had.

Will schedule in time to upgrade our CD environment and then revisit
upgrading to 5.x series.

Note: Just to clarify, the minimum supported Zk version for Apache
Curator is 3.5.x but debian is still on 3.4.13.