bugfix: don't include Bootstrap.css when open legacy DRM dialog in new UI
If Bootstrap.css is included in new UI, some components like buttons
will be broken. This could be triggered by opening the legacy DRM dialog
from either Summary page or the attachment list of old Search page.
So the fix is to add a flag which indicate whether Bootstrap.css is needed
or not.