refactor: remove unnecessary await modifier to act() call
Follow up from previous tidy up of this snippet in earlier commit. But
while here it was flagged in the PR that we're importing a different
`act` in SearchPage.test.tsx, so I decided we should use the one
specific to the RTL we're testing with.

As per feedback from PR #3373