refactor(front-end): focus the implementation of WizardHelper's getStringControlValue()
Really the function should only be concerned with a 'defined' value, as
`undefined` values should be handled before calling.

To support this, a few other areas were tweaked and tidied up. Maybe of
note is the isXyzArray() functions, which I realised due to the way
isHead() works actually narrows us right down to a NonEmptyArray.

Triggered through feedback in PR #3420