chore(deps): stop tika parsers transitive log4j dep
When we upgraded to Tika 2 it turns out that the microsoft parser in
standard parsers brings in log4j including an SLF4J binding. This causes
some of our logging to then fail, as it uses SLF4J and picks up the
Log4J2 binding where we only configure Log4J 1.2.x.

This was only new in the version we're currently working on and have not
yet released - 2021.2.