chore(deps): remove unused tomcat:jsp-api
This appears to be unused (based on testing by complete rebuild after
removal, and manual UI usage); most likely as there was also the
dependency for the newer org.apache.tomcat:tomcat-jsp-api. This newer
one was kept - and is in sync with all the other Tomcat deps due to a
variable.