chore(deps): remove references to asm
Not sure why this was explicitly included other than possibly when the
migration was done from the 'lib' directory to accommodate what is now a
transitive dependency.

Also, explicitly include a version meant the transitives had to be
excluded where as now that can be simplified and left to its own
devices.