build(deps): move struts dep to autotest only
Ideally I'd like to see struts gone, however it appears the IntegTester
uses struts for one of its UIs so we'll have to keep it for now.
However, rather than have it as a root dependency, makes sense to
isolate it to only where it's needed.

That said, it was kind of used but only for a Mock class in a unit test.
 Better to do this with Mockito instead.