build(deps): remove struts dependency (#3730)
Struts was really only used by the old IntegrationTester (which has since been re-written with scala/ps and now located at autotest/IntegTester). The only reason for the remaining dependency was to support a unit test which was using a RequestServlet mock provided by struts.

Better to do this with Mockito instead, and remove the dependency and the old (no longer built) code.