build(docker): upgrade to explicit openJDK base image
Previously we pinned to a hash to solve an issue, but then we no longer
new what version of openjdk we were on (turns out it was 8u212. So to
stay up to date we'll now use the tags again.

But also in update we had to navigate the changes in the underlying
linux - specifically the deprecation and removal of libav. So similar to
 our build etc we setup symbolic links to ffmpeg with is CLI compatible.

 Also did a minor change for the file name for the installer, just makes
  things simpler.