chore(deps): remove unused hudson xstream
Was going to convert to proper xstream, but first attempted removal.
Manual testing shows the admin console all works (including reading
configuration files - which was the only area that jumped out at me for
usage of XStream).

I'm assuming it probably gets it from elsewhere - as it does call into
common to get an XStream instance.