chore(deps): upgrade com.springsource.com.thoughtworks.xstream
Although still pretty old, best that can be done for now. Was hoping
this could be replaced with com.thoughtworks.xstream but that seems to
be not a like for like replacement - and things go bang.

Also note the removal of xmlpull, as xstream supplies that as a
transitive. So no need to explicitly manage.