chore(deps): upgrade Axis2 to 1.7.9
Couldn't go to 1.8.0 yet as there is axis2-jaxbri only has up to 1.7.9
(and I didn't want to mix).

Also removed the explicit axiom dependency as that's a transient of
Axis2 so I imagine it was only done due to the migration from ant/ivy
back in the day.

Lastly, a slight tweak to the circe dependencies in plugins.sbt so as to
 be consistent (and possibly make it easier for renovate to detect).