chore(deps): demarcate (softly) axis 1.4
Unfortunately we can't easily get rid of or upgrade Axis 1.4 due to the
very old library needed for SRW support - I can't find a newer version
(or its code) or an alternative. So the best I can do is call it out and
put it in its own block of dependencies. Hopefully then if in the future
we remove SRW support, then we can also easily remove the dependency.