chore(deps): partial revert of ansi-regex PR #3814
It's unclear how this really got into the lock file in this fashion,
however without builds fail in pristine environments (like openjdk
docker containers and GHA) - but still work on CodeBuild for some
(ominous) reason.

Also noted (as a red-herring) there is a slight bug in the GHA YAML
specification meaning we're not using the version of Node/NPM for the
functional tests - just what ever the default is.