refactor(search2): optimise number of attachment retrievals
Profiling revealed a lot of time was spent in search2 requests in
ifNotBroken due to both uses calling off to retrieve the same
attachment. Significant performance (both speed and DB calls)
improvements were possible if instead we re-use the attachment we
retrieve when we first check for broken attachment status rather than
throwing it away.