refactor: update CloudProviderService to drop the use of simpleDBA
To drop simpleDBA in CloudProviderService requires two tasks. One is to replace the
use of 'DBCacheBuilder' with 'ReplicatedCache' and the other one is to rewrite methods
that use the Monad of 'DB'. This means wherever these methods are used, changes are
requires as well to drop the 'DB' Monad there.

To improve the code readability and reduce duplication I put some common methods in a
helper file named 'CloudProviderHelper'. The most important method is probably the implicit
method that converts 'Entity' to 'CloudProviderInstance'.