feat(frontend): retrieve attachment details only on request
In institutions with lots of attachments the render of each search page
became super slow due to all the server side processing of attachments.
To avoid this unnecessary delay (considering most attachments will not
be looked at on the search results list) this code now only requests
attachment details if the attachment list is expanded (either by a user
or automatically due to display settings for the collection).

Work though remains, to address the thumbnails - which used the first
attachment.