fix: ensure each view of a webpage has an audit log
The audit logs for webpage attachments were missing because cache of the webpage
was used and no requests were sent to server. So we need to update the cache control
for webpage attachment so that a request is sent to check whether the cache is still
valid and this request will also trigger an audit log.