refactor: update the audit log content for viewing a webpage from summary page
The audit log content for a webpage can vary, depending on how it's viewed. For example,
if it's viewed from summary page, the MIME type(data3) is 'htmlpage'. But if it's viewed
directly through the file path, the MIME type is 'file:text/html'. To avoid confusion,
make sure the log content is consistent.