refactor: drop the feature of Cloud Search
This includes the removal of the whole feature as well as the tests and
REST endpoint. The Cloud Search setting is also dropped. If users saved
a Cloud search before and try to view the search again, a 503 error will
be thrown to indicate the feature is no longer supported.