refactor: create an APP level Context to provide current user details
Previously, if a component needs the current user details, it has to call the
REST endpoint again to get the details. This has caused a lot of unnecessary calls.
The solution is to use a top level Context to provide the user details so that any
component can access the details easily.