chore(deps): remove unused plugin dep axis2-jaxbri
It would appear this is not actually required, as the build appears to
work fine without it. However keeping it there, we're blocked from
upgrading to axis 1.8+ as there is no version of this in that release -
release notes say they renamed it, but they don't say what to. (Well,
they do but there seems to be a typo as they say it is renamed to the
same name.)