feat(front-end): add support for moderation sort orders
This change was ultimately about adding 'task_lastaction' and
'task_submitted' to SortOrderRunTypes. However, in doing so it came to
my attention that this type definition was sitting in OEQ.SearchSettings
rather than OEQ.Search. Further, the 'order' param in SearchParmsBase
was only defined as 'string' rather than the explicit type.

So I moved SortOrderRunTypes and SortOrder to OEQ.Search, added the
two new orders and then needed to update all the usages - hence the
updates to the other 13 files in react-front-end.

Lastly, it should be noted that I changed all the sort orders to
lowercase because on the server side, when it interprets the order param
first it forces them to lowercase and the apidocs are all focused on
lowercase. So we should be matching the expectation.

OEQ-1353