feat: support returning to New UI from Scrapbook legacy pages in Selection Session
To return to New UI with previous search criteria applied, we need to make sure the route
returned from server has query string `type` and `newUIStateId`. The approach is to add a
bookmarked field in the model class of RootMyResourcesSection so that this field will be
added to the path as a query string.