chore: exclude caniuse-lite from license checks
Suddenly this has started failing due to it's use of CC-BY-4.0, but it
has been using this for more than six years so I don't know what has
changed in the project's setup to trigger this. (I'm guessing it's the
updates in package-lock.json.)

Anyway, it's use of this license shouldn't be an issue for the project,
so I'll add it to the ignore list.