fix: address missing act() warnings in MyResourcesPage.test.tsx
The test was proving flaky, so when I noticed these it seemed likely to
be the reason. However although the warning says there's missing 'act'
statements, with the React Testing Library this typically means there's
something more fundamental wrong with the test(s) - as was the case here
.

Two main things was we were not properly waiting for the state changes
to complete; and, we had some state changes occurring in parallel due to
using Promise.all().