refactor(frontend): explicitly use DesktopDatePicker
Due to the way DatePicker does a `@media(pointer: fine)` media query to
switch between Mobile and Desktop variants, it was not possible to get
NewSearchPageTest to work on headless linux. It would always use the
Mobile Version for which the text fields are readonly. Previously we
didn't have a mobile solution anyway - as it's not a supported platform
for oEQ - so it makes sense to just go Desktop all the time.

Details of my last attempts were noted on GitHub at:
https://github.com/mui/mui-x/issues/4644#issuecomment-1343851120

MUI5 update

OEQ-1212