refactor: update how to import types in REST Module
In order to make sure TS files generated by io-ts-codegen have correct imports, we have to add a
rule for how to import type definitions in each REST Module file. The rule is a named import of a
type definition must explicitly be a type import.
For example,
```
 import type { A } from './X';
```
This also means one import clause has to be separated into two clauses if there are types imported
in the clause.