chore: make sure fp-ts is correctly resolved in react-front-end
After fp-ts was added in oeq-ts-rest-api, the use of fp-ts 'A.traverse' showed type errors.
This was probably because 'tsc' resolved 'fp-ts' to the wrong place. It should use the one
under react-front-end rather than the one under oeq-ts-rest-api. The real cause is still not
clear but a workaround is added now to help move work forward.