build(oeq-ts-rest-api): add missing ESLint deps
oeq-ts-rest-api maintains its own ESLint setup as the intention is to
split it out to its own module. However, it seemed to be missing a few
dependencies and instead was simply using those from the root of the
project. This only became obvious now that the codegen uses ESLint, and
on fresh CI builds of just this module the codegen would fail with
missing ESLint.