feat: add LTI 1.3 launch support
This includes the initial outline of support for LTI 1.3 launch. There
are a number of 'TODO's that need to be addressed, and some additional
error handling needed - to also conform to spec. Plus I would like to
add some tests (somehow). But this is a good point for initial code
review.

OEQ-1267