refactor: update the signatures of methods defined in LtiPlatformService
Previously, the Service layer tried to capture all potential exceptions. However, we
found that it could not capture all such as an exception thrown from `@transaction`.
So we had to add an additional error handling in the caller. To make code tidier and
more readable, we decided to do to error handling in the caller (in this case it's the
LtiPlatformResource). Hence, all the signatures have been updated as well as their
implementations.