feat: update io-ts-gen to support Map<K, V> and Set<T>
To automatically generate codec for these two types, we must generate a string representation
for the type first, and then use it with `customCombinator`.
The generated Codec requires `Ord` of the generic type to be provided. However, we do not care
the ordering at this stage. So we just do `OrdAsIts`.