feat: build a Deep linking response to complete a content selection
To complete a content selection, we need to send a Deep linking response back to the
LTI platform. According to the LTI 1.3 spec, an auto-submitted form is required to do
this. In OEQ, building this form and the script to submit the form varies, depening on
whether New UI or Old UI is turned on.