fix: support longer LTI 1.3 user ids
Although userids with moodle in LTI 1.3 were simple (short) numbers, it
turned out Canvas uses UUIDs. To support these the generation of the
value to store in tle_users.uuid was completely reworked to use a
combination of murmur3 hashing and base 64 representation of those hash
values. This results in typical values around the 38 character mark,
which happily fits in the VARCHAR(40) while still being suitably unique.

OEQ-1529