test: reduce jest concurrency
As the number and complexity of our react tests has increased, we're
finding more and more timeout issues. I believe this is due to CPU
thread starvation as by default Jest simply runs a number of workers to
match the number of CPU threads. And when the jest tests are running,
you can feel your whole machine is impacted.

By setting maxWorkers to 50% of available CPU threads all the tests have
started to pass again AND there's about a 25% speed up in test execution
time.