chore: remove the doc type declaration in 'checkstyle-config.xml'
After SBT is upgraded to v1.9.4, when executing SBT task 'checkJavaCodeStyle', there is error
about the DOCTYPE declaration in 'checkstyle-config.xml' -
DOCTYPE is disallowed when the feature "http://apache.org/xml/features/disallow-doctype-decl"
set to true.

For security concerns we probably do not want to change the value of 'disallow-doctype-decl', but
it's probably fine to remove the declaration.