chore: removed unused JSON Exporter / Remoting code
This code is unused, and the history of it has been lost - I looked in
the archives and all it told me was that it was part of the initial
import from the old repo in 2013. I attempted to trigger the code, and
got an error from JPF land highlighting that the extension point is
completely unused.