fix: create a JS var for the configured Collections and use it in CollectionSelector
What Collections are available to select in the New UI is also subject to the configuration
of Wizard control Resource selector. However, we do not currently have enough support to know
from which Collection of a Contribution the Selection session is launched, and as a result we
cannot query the Collection configuration. Luckily, we have back-end code that provides the
configured Collections, so we prepare a JS variable on server and provide it to the front-end.