feat: add REST endpoints for OIDC configuration
This commit includes a GET and PUT endpoint for retrieving and updating
the configuration. The annotation "JsonTypeInfo" and "JsonSubTypes" are
added to the abstract class in order to support the instantiation of a
concrete sub-class through Jackson.

The codec of IdentityProvide has also been refactored to help reduce the
complexity of using generic types in the Service.