feat: support saving OIDC configuration without providing secrets
As a security enhancement, sensitive values like API Client Secret are not
returned to client. As a result, when client sends a request back to update
the configuration, these values may be absent. In this case, the values of
an existing configuration will be reused. If the values are neither provided
or available in an existing configuration, errors will be returned for the missing
values.