refactor: make OAuth2Error sealed
The objective is simple, but it does require many changes in the LTI 1.3 integration
about how to support different types of errors. The end result is we have a wrapper
in LTI 1.3 land to provide the access to standard OAuth2 errors and a couple implicit
functions to do the type transformation. While doing this, I also simplified the function
doing the token verification by moving some logic out to the main Servlet.