feat: support different Kaltura embed style
This is the foundation for the Kaltura media embed code. Currently, the code is generated based
on the Auto Embed style, but it is not quite right and it does not work in LMS (e.g. Canvas).
As a result, we need to use the Iframe Embed style to generate the code.

While working on this, I also moved the Kaltura Embed component to a new directory dedicated for
the feature of embedding OEQ Attachment.